Issue-#58: Introduce "ConstraintBundle" interface to allow covering engines to access internal structure of constraints.
