+++ src/main/java/com/github/dakusui/jcunit/plugins/caengines/CoveringArrayEngine.java
@@ -3,14 +3,14 @@ package com.github.dakusui.jcunit.plugins.caengines;
 /**
  */
 
-import com.github.dakusui.jcunit.core.utils.BaseBuilder;
-import com.github.dakusui.jcunit.core.utils.Checks;
-import com.github.dakusui.jcunit.core.utils.Utils;
 import com.github.dakusui.jcunit.core.factor.FactorSpace;
 import com.github.dakusui.jcunit.core.factor.Factors;
 import com.github.dakusui.jcunit.core.tuples.Tuple;
+import com.github.dakusui.jcunit.core.utils.BaseBuilder;
+import com.github.dakusui.jcunit.core.utils.Checks;
+import com.github.dakusui.jcunit.core.utils.Utils;
 import com.github.dakusui.jcunit.plugins.Plugin;
-import com.github.dakusui.jcunit.plugins.constraints.ConstraintChecker;
+import com.github.dakusui.jcunit.plugins.constraints.ConstraintBundle;
 import com.github.dakusui.jcunit.runners.core.RunnerContext;
 import com.github.dakusui.jcunit.runners.standard.annotations.Generator;
 import com.github.dakusui.jcunit.runners.standard.annotations.Value;
@@ -32,19 +32,17 @@ public interface CoveringArrayEngine extends Plugin {
 
     private RunnerContext                        runnerContext;
     private Factors                              factors;
-    private ConstraintChecker                    constraintChecker;
     private Param.Resolver<S>                    resolver;
     private List<S>                              configArgsForEngine;
     private Class<? extends CoveringArrayEngine> engineClass;
 
     public Builder(
         RunnerContext runnerContext,
-        Factors factors, ConstraintChecker cm, Class<? extends CoveringArrayEngine> engineClass
+        Factors factors, Class<? extends CoveringArrayEngine> engineClass
     ) {
       this.runnerContext = checknotnull(runnerContext);
       this.factors = factors;
       this.engineClass = checknotnull(engineClass);
-      this.constraintChecker = checknotnull(cm);
     }
 
     public Builder setConfigArgsForEngine(List<S> parameters) {
@@ -57,22 +55,14 @@ public interface CoveringArrayEngine extends Plugin {
       return this;
     }
 
-
     public Factors getFactors() {
       return factors;
     }
 
-    public ConstraintChecker getConstraintChecker() {
-      return constraintChecker;
-    }
-
     public CoveringArrayEngine build() {
-      checknotnull(this.constraintChecker);
       checknotnull(this.engineClass);
       checknotnull(this.factors);
-      checknotnull(this.constraintChecker);
       checknotnull(this.runnerContext);
-      CoveringArrayEngine ret;
       Plugin.Factory<CoveringArrayEngine, S> factory;
       List<S> configArgsForEngine;
       if (this.resolver != null) {
@@ -94,8 +84,7 @@ public interface CoveringArrayEngine extends Plugin {
         );
         configArgsForEngine = Collections.emptyList();
       }
-      ret = factory.create(configArgsForEngine);
-      return ret;
+      return factory.create(configArgsForEngine);
     }
 
   }
@@ -112,7 +101,7 @@ public interface CoveringArrayEngine extends Plugin {
     final public CoveringArray generate(FactorSpace factorSpace) {
       Checks.checknotnull(factorSpace);
       return createCoveringArray(Utils.transform(
-          Utils.dedup(generate(factorSpace.factors, factorSpace.constraintChecker)),
+          Utils.dedup(generate(factorSpace.factors, ConstraintBundle.Factory.createFromConstraintChecker(factorSpace.constraintChecker))),
           new Utils.Form<Tuple, Tuple>() {
             @Override
             public Tuple apply(Tuple in) {
@@ -131,10 +120,10 @@ public interface CoveringArrayEngine extends Plugin {
      * generated by this object.
      * <p/>
      *
-     * @param factors    factors from which tuples will be generated.
-     * @param constraintChecker constraint on tuple generation.
+     * @param factors           factors from which tuples will be generated.
+     * @param constraintBundle  A bundle of constraints used on tuple generation.
      */
-    abstract protected List<Tuple> generate(Factors factors, ConstraintChecker constraintChecker);
+    abstract protected List<Tuple> generate(Factors factors, ConstraintBundle constraintBundle);
   }
 
   class FromAnnotation implements BaseBuilder<CoveringArrayEngine> {
