+++ src/main/java/com/github/dakusui/jcunit/plugins/caengines/StandardCoveringArrayEngine.java
@@ -1,9 +1,9 @@
 package com.github.dakusui.jcunit.plugins.caengines;
 
-import com.github.dakusui.jcunit.core.utils.Checks;
 import com.github.dakusui.jcunit.core.factor.Factors;
 import com.github.dakusui.jcunit.core.tuples.Tuple;
-import com.github.dakusui.jcunit.plugins.constraints.ConstraintChecker;
+import com.github.dakusui.jcunit.core.utils.Checks;
+import com.github.dakusui.jcunit.plugins.constraints.ConstraintBundle;
 
 import java.util.List;
 
@@ -16,12 +16,12 @@ public class StandardCoveringArrayEngine extends CoveringArrayEngine.Base {
   }
 
   @Override
-  protected List<Tuple> generate(Factors factors, ConstraintChecker constraintChecker) {
+  protected List<Tuple> generate(Factors factors, ConstraintBundle constraintBundle) {
     Checks.checknotnull(factors);
-    Checks.checknotnull(constraintChecker);
+    Checks.checknotnull(constraintBundle);
     if (factors.size() < 2 || this.strength < 2) {
-      return new SimpleCoveringArrayEngine().generate(factors, constraintChecker);
+      return new SimpleCoveringArrayEngine().generate(factors, constraintBundle);
     }
-    return new Ipo2CoveringArrayEngine(this.strength).generate(factors, constraintChecker);
+    return new Ipo2CoveringArrayEngine(this.strength).generate(factors, constraintBundle);
   }
 }
