+++ src/main/java/com/github/dakusui/jcunit/utils/CoveringArrayEngines.java
@@ -4,7 +4,6 @@ import com.github.dakusui.jcunit.core.factor.Factors;
 import com.github.dakusui.jcunit.plugins.PluginUtils;
 import com.github.dakusui.jcunit.plugins.caengines.CoveringArrayEngine;
 import com.github.dakusui.jcunit.plugins.caengines.Ipo2CoveringArrayEngine;
-import com.github.dakusui.jcunit.plugins.constraints.ConstraintChecker;
 import com.github.dakusui.jcunit.runners.core.RunnerContext;
 
 import java.util.Arrays;
@@ -15,31 +14,17 @@ import java.util.Arrays;
 public enum CoveringArrayEngines {
   ;
 
-  private static CoveringArrayEngine.Builder<String[]> createSimpleBuilder(
+  public static CoveringArrayEngine.Builder<String[]> createSimpleBuilder(
       Factors factors,
-      ConstraintChecker constraintChecker,
       Class<? extends CoveringArrayEngine> engineClass,
       String[]... configArgsForEngine
   ) {
     return new CoveringArrayEngine.Builder<String[]>(
         RunnerContext.DUMMY,
         factors,
-        constraintChecker,
         engineClass
     ).setResolver(PluginUtils.StringArrayResolver.INSTANCE).setConfigArgsForEngine(Arrays.asList(configArgsForEngine));
   }
-
-  public static CoveringArrayEngine.Builder<String[]> createSimpleBuilder(
-      Factors factors,
-      Class<? extends CoveringArrayEngine> engineClass,
-      String[]... arguments) {
-    return createSimpleBuilder(
-        factors,
-        ConstraintChecker.DEFAULT_CONSTRAINT_CHECKER,
-        engineClass,
-        arguments);
-  }
-
 
   public static CoveringArrayEngine.Builder<String[]> createSimpleBuilder(Factors factors) {
     return createSimpleBuilder(factors, Ipo2CoveringArrayEngine.class);
