+++ src/test/java/com/github/dakusui/jcunit/tests/bugfixes/geophile/NullCoveringArrayEngine.java
@@ -3,14 +3,14 @@ package com.github.dakusui.jcunit.tests.bugfixes.geophile;
 import com.github.dakusui.jcunit.core.factor.Factors;
 import com.github.dakusui.jcunit.core.tuples.Tuple;
 import com.github.dakusui.jcunit.plugins.caengines.CoveringArrayEngine;
-import com.github.dakusui.jcunit.plugins.constraints.ConstraintChecker;
+import com.github.dakusui.jcunit.plugins.constraints.ConstraintBundle;
 
 import java.util.Collections;
 import java.util.List;
 
 public class NullCoveringArrayEngine extends CoveringArrayEngine.Base {
   @Override
-  protected List<Tuple> generate(Factors factors, ConstraintChecker constraintChecker) {
+  protected List<Tuple> generate(Factors factors, ConstraintBundle constraintBundle) {
     return Collections.emptyList();
   }
 }
